packages/deterministic_lv_model/pyproject.toml
[project]
name = "deterministic-lv-model"
version = "0.1.0"
description = "Deterministic Lotka-Volterra model"
readme = "README.md"
authors = [
    { name = "Eric Mooring", email = "pgv5@cdc.gov" },
]
requires-python = ">=3.12"
dependencies = ["numpy>=2.3.4", "scipy>=1.11", "matplotlib>=3.8", "cfa-mrp"]

[project.scripts]
deterministic_lv_model = "deterministic_lv_model.deterministic_lv_model:main"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"
